Project Import | Preselect only first conflicting extension according to load order
