+++ modules/project/import-core/src/sap/commerce/toolset/project/descriptor/impl/YCommonWebSubModuleDescriptor.kt
