+++ modules/project/import-core/src/sap/commerce/toolset/project/descriptor/impl/YCoreExtModuleDescriptor.kt
@@ -28,5 +28,5 @@ class YCoreExtModuleDescriptor(
 ) : YPlatformExtModuleDescriptor(moduleRootPath, extensionInfo) {
 
     override var importStatus = ModuleDescriptorImportStatus.MANDATORY
-    override fun initDependencies(moduleDescriptors: Map<String, ModuleDescriptor>) = emptySet<String>()
+    override fun initDependencies(moduleDescriptors: Map<String, Collection<ModuleDescriptor>>) = emptySet<String>()
 }
